Refactor: moved files to research, basic cython idea
[*] Added very basic cython target (helpers)
    - Report really isn't very promising (pretty much the whole file
    was interacting very heavily with python)
    - TODO: ensure helpers is more portable to cython

[*] Created seperate helpers file in scripts because ^^^

[*] Cleaned up some files that shouldn't be in allocator (at the moment)

[*] Added .gitignore to ignore cython build artefacts

[*] Misc:
    [*] Fix up imports
