+++ Src/Allocator/.gitignore
@@ -1,0 +1,3 @@
+*.c
+build/
+*.html
