+++ Src/Allocator/Interpreter/requirements.txt
@@ -10,6 +10,7 @@ codespell==2.4.1
 comm==0.2.2
 contourpy==1.3.1
 cycler==0.12.1
+Cython==3.1.1
 debugpy==1.8.13
 decorator==5.2.1
 distlib==0.3.9
