+++ Src/RTL/Scripts/i2s_mono_stereo_wrapper.py
@@ -16,7 +16,7 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(
 from consts import MONO_STEREO_WRAPPER_PREFIX, I2S_DUPLICATE_REGISTER_HEADER_PATH,\
 I2S_DUPLICATE_REGISTER_PATH
 from Allocator.Interpreter.consts import LOGGER
-from Allocator.Interpreter.helpers import str2bool
+from argparse_helpers import str2bool
 
 
 def main() -> None:
